Update accounts_service.ts
If you use name instead of username, it will not register a user because Accounts.createUser is looking for username and not name.  Also added email because it is easier for code completion and also gives all fields.  Plus, if you ever wanted to use Accounts.forgotPassword then there would be no way for user to reset password without getting you or Admin to manually reset it.  email is more important for that reason than username.  That is why I set username as optional.